requirements-freeze.txt
a2wsgi==1.10.7
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
bentoml==1.3.15
billiard==4.2.1
cattrs==23.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
deepmerge==2.0
Deprecated==1.2.15
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.4
dvc==3.53.2
dvc-data==3.15.2
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.16.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
funcy==2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grandalf==0.8
grpcio==1.67.1
gto==1.7.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.0
httpx-ws==0.6.2
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
iterative-telemetry==0.0.9
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
omegaconf==2.3.0
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.11
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
pip==22.0.2
pip-requirements-parser==32.0.1
platformdirs==3.11.0
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.19
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
schema==0.7.7
scmrepo==3.3.8
semver==3.0.2
setuptools==59.6.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simple-di==0.1.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqltrie==0.11.1
starlette==0.41.3
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tomli_w==1.1.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.6
uvicorn==0.32.1
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.0.0
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zc.lockfile==3.0.post1
zipp==3.21.0
